+++ models/staging/stg_yellow_tripdata.sql
@@ -4,14 +4,14 @@
 with tripdata as 
 (
   select *,
-    row_number() over(partition by CAST(vendorid as integer), tpep_pickup_datetime) as rn
+    row_number() over(partition by CAST(vendorid as numeric), tpep_pickup_datetime) as rn
   from {{ source('staging','yellow_tripdata') }}
   where vendorid is not null 
 )
 select
    -- identifiers
     {{ dbt_utils.surrogate_key(['vendorid', 'tpep_pickup_datetime']) }} as tripid,
-    cast(vendorid as integer) as vendorid,
+    cast(vendorid as numeric) as vendorid,
     cast(ratecodeid as integer) as ratecodeid,
     cast(pulocationid as integer) as  pickup_locationid,
     cast(dolocationid as integer) as dropoff_locationid,
